修改日志文件存放目录。修复收据用水量显示异常问题。添加用户时增加存在时更新以保证程序不出错

+++ rental-application/src/main/java/site/minnan/rental/infrastructure/utils/ReceiptUtils.java
@@ -271,7 +271,7 @@ public class ReceiptUtils {
         //水费
         context.setVariable("waterStart", bill.getWaterStart().intValue());
         context.setVariable("waterEnd", bill.getWaterEnd().intValue());
-        context.setVariable("waterUsage", bill.getWaterUsage());
+        context.setVariable("waterUsage", bill.getWaterUsage().intValue());
         context.setVariable("waterPrice", price.getWaterPrice().intValue());
         String[] waterCharge = splitNumber(bill.getWaterCharge().intValue());
         context.setVariable("waterCharge", waterCharge);
